+++ client/src/main.jsx
@@ -1,12 +1,14 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App'
+
 import {
   createBrowserRouter,
   RouterProvider,
 } from "react-router-dom";
 import './index.css'
 import Root from "./routes/root";
+import FormNewIndividual from './components/FormNewIndividual';
 
 const router = createBrowserRouter([
   {
@@ -16,6 +18,10 @@ const router = createBrowserRouter([
   {
     path: "/App",
     element: <App />,
+  },
+  {
+    path: "/new-individual",
+    element: <FormNewIndividual />,
   }
 ]);
